Move the layout to the high order folder as it is a high order component since it is only there to wrap other components
