+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder';
-import Layout from './components/Layout/Layout';
+import Layout from './hoc/Layout/Layout';
 
 class App extends Component {
   render() {
